+++ projects/js-packages/publicize-components/index.ts
@@ -3,6 +3,9 @@
 //store for the components, but at the moment they're tied.
 import './src/social-store';
 
+// Ensure that module augmentation is applied
+export type {} from './src/types';
+
 export { default as Connection } from './src/components/connection';
 export { default as Form } from './src/components/form';
 export { default as SocialPreviewsModal } from './src/components/social-previews/modal';