+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/breve/controls.js
@@ -35,7 +35,8 @@ export const useInit = init => {
 
 const Controls = ( { blocks, disabledFeatures } ) => {
 	const [ gradeLevel, setGradeLevel ] = useState( null );
-	const { toggleFeature, toggleProofread } = useDispatch( 'jetpack/ai-breve' );
+	const { toggleFeature, toggleProofread, setPopoverHover, setHighlightHover, setPopoverAnchor } =
+		useDispatch( 'jetpack/ai-breve' );
 	const { tracks } = useAnalytics();
 
 	const isProofreadEnabled = useSelect(
@@ -82,6 +83,19 @@ const Controls = ( { blocks, disabledFeatures } ) => {
 	// Update the grade level immediately on first load.
 	useInit( updateGradeLevel );
 
+	// Disable the popover when proofread or a feature is disabled.
+	useEffect( () => {
+		setPopoverHover( false );
+		setHighlightHover( false );
+		setPopoverAnchor( { target: null, virtual: null } );
+	}, [
+		setPopoverHover,
+		setHighlightHover,
+		setPopoverAnchor,
+		isProofreadEnabled,
+		disabledFeatures,
+	] );
+
 	return (
 		<div className="jetpack-ai-proofread">
 			<p> { __( 'Improve your writing with AI.', 'jetpack' ) }</p>
@@ -120,7 +134,7 @@ const Controls = ( { blocks, disabledFeatures } ) => {
 							<CheckboxControl
 								className={ isProofreadEnabled ? '' : 'is-disabled' }
 								disabled={ ! isProofreadEnabled }
-								data-type={ feature.config.name }
+								data-breve-type={ feature.config.name }
 								key={ feature.config.name }
 								label={ feature.config.title }
 								checked={ ! disabledFeatures.includes( feature.config.name ) }